Refactor of 'smart' example packaging, and introduction of abstract federate 'helper' classes to reduce federate code.
