+++ src/java/main/gov/nist/ucef/hla/example/smart/federates/SmartPingFederate.java
@@ -21,19 +21,17 @@
  * NOT HAVE ANY OBLIGATION TO PROVIDE MAINTENANCE, SUPPORT, UPDATES, ENHANCEMENTS, OR
  * MODIFICATIONS.
  */
-package gov.nist.ucef.hla.example.smart;
+package gov.nist.ucef.hla.example.smart.federates;
 
 import gov.nist.ucef.hla.base.FederateConfiguration;
-import gov.nist.ucef.hla.base.HLAInteraction;
-import gov.nist.ucef.hla.base.HLAObject;
 import gov.nist.ucef.hla.base.UCEFException;
 import gov.nist.ucef.hla.base.UCEFSyncPoint;
+import gov.nist.ucef.hla.example.smart.helpers._SmartPingFederate;
 import gov.nist.ucef.hla.example.smart.interactions.Ping;
 import gov.nist.ucef.hla.example.smart.interactions.Pong;
 import gov.nist.ucef.hla.example.smart.reflections.Player;
 import gov.nist.ucef.hla.example.util.Constants;
 import gov.nist.ucef.hla.example.util.FileUtils;
-import gov.nist.ucef.hla.ucef.NoOpFederate;
 import gov.nist.ucef.hla.ucef.SimEnd;
 import gov.nist.ucef.hla.ucef.SimPause;
 import gov.nist.ucef.hla.ucef.SimResume;
@@ -50,22 +48,23 @@ import gov.nist.ucef.hla.ucef.SimResume;
  * 
  * Example federate for testing
  */
-public class SmartPongFederate extends NoOpFederate
+public class SmartPingFederate extends _SmartPingFederate
 {
 	//----------------------------------------------------------
 	//                   STATIC VARIABLES
 	//----------------------------------------------------------
+	private static String[] PLAYER_NAMES= {"Alice Ping", "Bob Ping", "Carol Ping", "David Ping"};
 
 	//----------------------------------------------------------
 	//                   INSTANCE VARIABLES
 	//----------------------------------------------------------
-	private char letter;
+	private int count;
 	private Player player;
 
 	//----------------------------------------------------------
 	//                      CONSTRUCTORS
 	//----------------------------------------------------------
-	public SmartPongFederate( String[] args )
+	public SmartPingFederate( String[] args )
 	{
 		super();
 	}
@@ -86,57 +85,51 @@ public class SmartPongFederate extends NoOpFederate
 	@Override
 	public void beforeFirstStep()
 	{
-		this.letter = 'a';
-		this.player = new Player( rtiamb, "PongPlayer" );
+		this.count =  0;
+		this.player = new Player( rtiamb, "PingPlayer" );
 	}
 
 	@Override
 	public boolean step( double currentTime )
 	{
 		// here we end out our interaction and attribute update
-		sendInteraction( new Pong( rtiamb, this.letter ) );
+		sendInteraction( new Ping( rtiamb, this.count ) );
 		updateAttributeValues( this.player );
 		// update the values
-		this.letter++;
-		String nextPlayerName = this.player.name() + this.letter;
+		this.count++;
+		String nextPlayerName = PLAYER_NAMES[this.count % PLAYER_NAMES.length];
 		this.player.name( nextPlayerName );
 		// keep going until time 10.0
 		return (currentTime < 10.0);
 	}
 
 	////////////////////////////////////////////////////////////////////////////////////////////
-	/////////////////////////////////// RTI Callback Methods ///////////////////////////////////
+	////////////////////////// Interaction/Reflection Handler Callbacks ////////////////////////
 	////////////////////////////////////////////////////////////////////////////////////////////
+	/**
+	 * Handle receipt of a {@link Pong}
+	 * 
+	 * @param pong the interaction to handle
+	 */
 	@Override
-	public void receiveInteraction( HLAInteraction hlaInteraction )
+	protected void receivePongInteraction( Pong pong )
 	{
-		if( rtiamb.isOfKind( hlaInteraction, Ping.interactionName() ) )
-		{
-			receivePingInteraction( new Ping( hlaInteraction ) );
-		}
-		else
-		{
-			// this is unexpected - we shouldn't receive any thing we didn't subscribe to
-			System.err.println( String.format( "Received an unexpected interaction of type '%s'",
-			                                   rtiamb.getInteractionClassName( hlaInteraction ) ) );
-		}
+		System.out.println( String.format( "Received Pong interaction - letter is '%s'.",
+		                                   pong.letter() ) );
 	}
 
+	/**
+	 * Handle receipt of a {@link Player}
+	 * 
+	 * @param player the object to handle
+	 */
 	@Override
-	public void receiveAttributeReflection( HLAObject hlaObject ) 
-	{ 
-		if( rtiamb.isOfKind( hlaObject, Player.objectClassName() ) )
-		{
-			receivePlayerUpdate( new Player( hlaObject ) );
-		}
-		else
-		{
-			// this is unexpected - we shouldn't receive any thing we didn't subscribe to
-			System.err.println( String.format( "Received an unexpected attribute reflection of type '%s'",
-			                                   rtiamb.getObjectClassName( hlaObject ) ) );
-		}
+	protected void receivePlayerUpdate( Player player )
+	{
+		System.out.println( String.format( "Received Player update - name is %s",
+		                                   player.name() ) );
 	}
-
+	
 	////////////////////////////////////////////////////////////////////////////////////////////
 	////////////////////// UCEF Simulation Control Interaction Callbacks ///////////////////////
 	////////////////////////////////////////////////////////////////////////////////////////////
@@ -158,31 +151,6 @@ public class SmartPongFederate extends NoOpFederate
 		System.out.println( "SimEnd signal received. Simulation will be terminated..." );
 	}
 
-	////////////////////////////////////////////////////////////////////////////////////////////
-	///////////////////////////////// Internal Utility Methods /////////////////////////////////
-	////////////////////////////////////////////////////////////////////////////////////////////
-	/**
-	 * Handle receipt of a {@link Ping}
-	 * 
-	 * @param ping the interaction to handle
-	 */
-	private void receivePingInteraction( Ping ping )
-	{
-		System.out.println( String.format( "Received Ping interaction - count is %s",
-		                                   ping.count() ) );
-	}
-
-	/**
-	 * Handle receipt of a {@link Player}
-	 * 
-	 * @param player the object to handle
-	 */
-	private void receivePlayerUpdate( Player player )
-	{
-		System.out.println( String.format( "Received Player update - name is %s",
-		                                   player.name() ) );
-	}
-	
 	//----------------------------------------------------------
 	//                     STATIC METHODS
 	//----------------------------------------------------------
@@ -193,16 +161,16 @@ public class SmartPongFederate extends NoOpFederate
 	 */
 	private static FederateConfiguration makeConfig()
 	{
-		FederateConfiguration config = new FederateConfiguration( "Pong",                 // name
-		                                                          "PongFederate",         // type
+		FederateConfiguration config = new FederateConfiguration( "Ping",                 // name
+		                                                          "PingFederate",         // type
 		                                                          "PingPongFederation" ); // execution
 
 		// set up lists of objects/attributes to be published and subscribed to
 		config.addPublishedAttributes( Player.objectClassName(), Player.attributeNames() );
 		config.addSubscribedAttributes( Player.objectClassName(), Player.attributeNames() );
 		// set up lists of interactions to be published and subscribed to
-		config.addPublishedInteraction( Pong.interactionName() );
-		config.addSubscribedInteraction( Ping.interactionName() );
+		config.addPublishedInteraction( Ping.interactionName() );
+		config.addSubscribedInteraction( Pong.interactionName() );
 		// subscribed UCEF simulation control interactions
 		config.addSubscribedInteractions( SimPause.interactionName(), SimResume.interactionName(),
 		                                  SimEnd.interactionName() );
@@ -236,20 +204,21 @@ public class SmartPongFederate extends NoOpFederate
 	{
 		System.out.println( Constants.UCEF_LOGO );
 		System.out.println();
-		System.out.println( "	______   ____   ____    ____  " );
-		System.out.println( "	\\____ \\ /  _ \\ /    \\  / ___\\" );
-		System.out.println( "	|  |_> >  <_> )   |  \\/ /_/  >" );
-		System.out.println( "	|   __/ \\____/|___|  /\\___  /" );
-		System.out.println( "	|__|               \\//_____/" );
-		System.out.println( "	     Smart Pong Federate" );
+		System.out.println( "	       .__                " );
+		System.out.println( "	______ |__| ____    ____" );
+		System.out.println( "	\\____ \\|  |/    \\  / ___\\" );
+		System.out.println( "	|  |_> >  |   |  \\/ /_/  >" );
+		System.out.println( "	|   __/|__|___|  /\\___  /" );
+		System.out.println( "	|__|           \\//_____/" );
+		System.out.println( "	   Smart Ping Federate" );
 		System.out.println();
-		System.out.println( "Sends 'Pong' interactions.");
-		System.out.println( "Receives 'Ping' interactions.");
+		System.out.println( "Sends 'Ping' interactions.");
+		System.out.println( "Receives 'Pong' interactions.");
 		System.out.println();
 
 		try
 		{
-			new SmartPongFederate( args ).runFederate( makeConfig() );
+			new SmartPingFederate( args ).runFederate( makeConfig() );
 		}
 		catch( Exception e )
 		{
