+++ src/java/main/gov/nist/ucef/hla/example/smart/reflections/Player.java
@@ -74,7 +74,7 @@ public class Player extends HLAObject
 	/////////////////////////////// Accessor and Mutator Methods ///////////////////////////////
 	////////////////////////////////////////////////////////////////////////////////////////////
 
-	public void isNamePresent( String name )
+	public void isNamePresent()
 	{
 		isPresent( ATTRIBUTE_KEY_NAME );
 	}
