+++ src/java/main/gov/nist/ucef/hla/ucef/SimResume.java
@@ -32,7 +32,7 @@ public class SimResume extends HLAInteraction
 	//                    STATIC VARIABLES
 	//----------------------------------------------------------
 	// HLA identifier of this type of interaction - must match FOM definition 
-	private static final String INTERACTION_NAME = "HLAInteractionRoot.C2WInteractionRoot.SimResume";
+	private static final String INTERACTION_NAME = "HLAInteractionRoot.C2WInteractionRoot.SimulationControl.SimResume";
 	
 	// interaction parameters and types
 	// ...none...
